code modularised into own file
relates to #108
